Change floats to double for improved consistency
- Also fixes Date being off by a few seconds

+++ xchange-cavirtex/src/main/java/com/xeiam/xchange/virtex/VirtExUtils.java
@@ -48,7 +48,7 @@ public class VirtExUtils {
    * @param price
    * @return
    */
-  public static BigMoney getPrice(String currency, float price) {
+  public static BigMoney getPrice(String currency, double price) {
 
     return MoneyUtils.parseFiat(currency + " " + new BigDecimal(price));
   }
